pyproject.toml
[tool.poetry]
name = "zuai"
version = "0.1.0"
description = "ZuAI Backend Engineer Evaluation Task"
authors = ["Satyarth Raghuvanshi <lightyagami.6320@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
fastapi = "^0.114.1"
uvicorn = "^0.30.6"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
google-generativeai = "^0.8.3"
motor = "^3.6.0"
setuptools = "^75.2.0"
redis = "^5.1.1"
aiofiles = "^24.1.0"
python-multipart = "^0.0.12"
pytest-asyncio = "^0.24.0"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
fastapi-limiter = "^0.1.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.7.0"
autopep8 = "^2.0.4"
pre-commit = "^3.4.0"
fpdf = "^1.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
